+++ lib/pages/settings/flags_page.dart
@@ -45,9 +45,8 @@ class _FlagsPageState extends State<FlagsPage> {
             privateFlag,
             enableNotificationsFlag,
             showBrightSchemeFlag,
-            allowInvalidCertFlag,
-            enableSyncInboxFlag,
-            enableSyncOutboxFlag,
+            // TODO: is this required? Remove
+            // allowInvalidCertFlag,
           };
 
           final filteredItems =