+++ lib/utils/consts.dart
@@ -2,5 +2,4 @@ const privateFlag = 'private';
 const enableNotificationsFlag = 'enable_notifications';
 const showBrightSchemeFlag = 'show_bright_scheme';
 const allowInvalidCertFlag = 'allow_invalid_cert';
-const enableSyncInboxFlag = 'enable_sync_inbox';
-const enableSyncOutboxFlag = 'enable_sync_outbox';
+const enableSyncFlag = 'enable_sync';