+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed:
 - Remove unused theme config widget
+- Hide config flag: allow_invalid_cert
 
 ## [0.8.250] - 2023-01-24
 ### Changed: